pm_runtime: hide platform specific structure
Hides platform specific structure in cAVS platform header.
