+++ src/include/sof/lib/pm_memory.h
@@ -14,24 +14,8 @@
 #ifndef __SOF_LIB_PM_MEMORY_H__
 #define __SOF_LIB_PM_MEMORY_H__
 
-#include <sof/spinlock.h>
-#include <sof/trace/trace.h>
-#include <user/trace.h>
+#include <platform/lib/pm_memory.h>
 #include <stdint.h>
-
-/** \brief Memory banks pm masks data. */
-struct ebb_data {
-	uint32_t current_mask0;
-	uint32_t current_mask1;
-	uint32_t new_mask0;
-	uint32_t new_mask1;
-	uint32_t change_mask0;
-	uint32_t change_mask1;
-	uint32_t start_bank_id;
-	uint32_t ending_bank_id;
-	uint32_t start_bank_id_high;
-	uint32_t ending_bank_id_high;
-};
 
 /**
  * \brief Set power gating of memory banks in the address range
@@ -43,7 +27,7 @@ struct ebb_data {
  * \param[in] size Size of memory to manage.
  * \param[in] enabled Boolean deciding banks desired state (1 powered 0 gated).
  */
-void set_power_gate_for_memory_address_range(void *ptr,
-					     uint32_t size, uint32_t enabled);
+void set_power_gate_for_memory_address_range(void *ptr, uint32_t size,
+					     uint32_t enabled);
 
 #endif /* __SOF_LIB_PM_MEMORY_H__ */
