+++ src/platform/icelake/include/platform/lib/pm_memory.h
@@ -1,0 +1,27 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ *
+ * Copyright(c) 2020 Intel Corporation. All rights reserved.
+ *
+ * Author: Tomasz Lauda <tomasz.lauda@linux.intel.com>
+ */
+
+/**
+ * \file platform/lib/pm_memory.h
+ * \brief Memory power management header file for Icelake
+ * \author Tomasz Lauda <tomasz.lauda@linux.intel.com>
+ */
+
+#ifdef __SOF_LIB_PM_MEMORY_H__
+
+#ifndef __PLATFORM_LIB_PM_MEMORY_H__
+#define __PLATFORM_LIB_PM_MEMORY_H__
+
+#include <cavs/lib/pm_memory.h>
+
+#endif /* __PLATFORM_LIB_PM_MEMORY_H__ */
+
+#else
+
+#error "Do not include outside of sof/lib/pm_memory.h"
+
+#endif /* __SOF_LIB_PM_MEMORY_H__ */
